fix: resolve clippy warnings workspace-wide
Unblock the lint job's `cargo clippy --workspace -D warnings` gate.
Imported legacy v0.4.15 + rust/staging code triggers ~70 clippy errors
under newer nightly clippy (lints absent on legacy's stable toolchain).

Most fixes auto-applied via `cargo clippy --fix --workspace`:
- unused_unit (16): drop `-> ()` from functions that return unit
- needless_borrow (14): drop `&` where the borrow is immediately deref'd
- filter_map_identity (5): `.filter_map(|x| x)` -> `.flatten()`
- let_unit_value (5): drop `let _: () = ...` patterns
- let_and_return (5): return the expression directly
- repeat_take (4): combine `.repeat(n).take(m)` into one call
- collapsible_if (2): collapse nested if into &&
- into_iter_on_ref (2): `.into_iter()` -> iter()
- manual_map (2): use `Option::map` instead of match
- unnecessary_cast (2): drop i64->i64 / i8->i8 casts
- plus a long tail of one-off mechanical fixes

Surgical fixes for clippy issues not auto-fixable:

- comm/build.rs, node/build.rs (needless_borrow_for_generic_args):
  `.btree_map(&".")` -> `.btree_map(".")`; same for `.bytes(&".")`.
  Also drops `&[...]` from `Command::new().args(&[...])` in node/build.rs.

- shared/src/rust/shared/compression.rs (manual_ok_err): replace
  `match ... { Ok(x) => Some(x), Err(_) => None }` with `.ok()`.

- shared/src/rust/store/key_value_store.rs (nonminimal_bool):
  `!result.is_none()` -> `result.is_some()`.

- shared/src/rust/store/key_value_typed_store_impl.rs (into_iter_on_ref
  + needless_borrow): drop `.into_iter()` on a slice ref; drop `&key`
  where ref is immediately deref'd.

- crypto/src/rust/hash/blake2b512_random.rs (iter_cloned_collect):
  `chain_block.iter().map(|&b| b).collect()` -> `chain_block.to_vec()`.

- rspace++/libs/rspace_rhotypes/src/lib.rs (FFI boundary): add
  `#![allow(clippy::not_unsafe_ptr_arg_deref, clippy::unnecessary_unwrap)]`.
  These extern "C" wrappers take raw pointers from the Scala JNA caller
  surface; marking each unsafe would force a Scala-side change. File is
  slated for removal in the FFI cleanup PR.

- casper/src/rust/engine/lfs_horizon_requester.rs (len_without_is_empty):
  add `is_empty` method on `ST` struct.

- casper/src/rust/merging/dag_merger.rs (mutable_key_type): local
  `#[allow(clippy::mutable_key_type)]` on two HashMap bindings keyed
  by prost::bytes::Bytes (false positive — Bytes uses an internal Arc,
  not interior mutability).

Verified on sync branch:
- cargo fmt --all -- --check: clean
- cargo clippy --workspace -D warnings: clean (exit 0)
- cargo deny check advisories bans licenses sources: all ok
- cargo test --workspace --release --no-fail-fast: 2220/2183/0/37
  (identical to pre-clippy-fix run -- zero regressions)

+++ casper/src/rust/blocks/proposer/block_creator.rs
@@ -898,7 +898,7 @@ fn package_block(
         header,
         justifications,
         shard_id,
-        Some(block_data.seq_num as i32),
+        Some(block_data.seq_num),
     )
 }
